Changed the way the user-agents and the mongodb connection is handled
First big change is that the user_agents.txt file is first read and an array of all user agents is build up. Then this array is passed once to db.collection(...).insert() function which makes it much faster compared to the previous state. Mongodb connection has been updated to the MongoClient class API